Set data type of constraint array to float

+++ constraint_tightening/constraint_tightening.py
@@ -5,10 +5,10 @@ from scipy import stats
 
 class ConstraintTightening:
     def __init__(self, G_u, g_u, G_x, g_x, risk_factor=0.95):
-        self.G_u = np.array(G_u)
-        self.g_u = np.array(g_u)
-        self.G_x = np.array(G_x)
-        self.g_x = np.array(g_x)
+        self.G_u = np.array(G_u,dtype=float)
+        self.g_u = np.array(g_u,dtype=float)
+        self.G_x = np.array(G_x,dtype=float)
+        self.g_x = np.array(g_x,dtype=float)
 
         # Initialize pseudo constraints
         self.G_v = self.G_u.copy()
